backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==45.0.7
defusedxml==0.7.1
Django==4.2.24
django-filter==24.3
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
gunicorn==23.0.0
idna==3.10
oauthlib==3.3.1
pillow==11.3.0
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
python3-openid==3.2.0
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
drf-extra-fields==3.4.1

django-storages[azure]==1.14.4

opencensus-ext-azure==1.1.13
opencensus-ext-django==0.8.0
opencensus-ext-logging==0.1.1

django-cors-headers==4.3.1

flake8==7.0.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.0.0

